+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/IRelationUpsertStrategyService.kt
@@ -31,6 +31,7 @@ interface IRelationUpsertStrategyService {
     data class UpsertRequest(
         val source: LegalEntityDb,
         val target: LegalEntityDb,
-        val validityPeriods: Collection<RelationValidityPeriodDb>
+        val validityPeriods: Collection<RelationValidityPeriodDb>,
+        val existingRelation: RelationDb?
     )
 }